added detection for git submodules
check for mode 57344 on a directory to indicate it is actually a git submodule. GitTree::listRecursive() uses is_submodule flag to know not to traverse that directory. :submodule is appended to the string name for that directory to indicate it is a submodule and not part of the current repo